+++ backend/src/main/java/com/example/Xavier/myapplication/backend/User.java
@@ -15,7 +15,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String username;
     private String password;
     private byte [] image;
@@ -32,16 +32,16 @@ public class User {
         this.image=image;
     }
 
-    public User(int id, String username, String password, byte [] image){
+    public User(Long id, String username, String password, byte [] image){
         this.id=id;
         this.username=username;
         this.password=password;
         this.image=image;
     }
 
-    public int getId() { return id; }
+    public Long getId() { return id; }
 
-    public void setId(int id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
 
     public byte[] getImage() { return image; }
